Prevent destroyed slider from running its resize handler.
Under some circumstances I see error reports with the latest jqRangeSlider showing a crash in _initWidth (invoked by resize) because this.container is null.  This should not happen because container is only nulled during destroy, which also removes the resize handler.  I guess that the resize event is already in progress at some level, and continues regardless.  The stack report does go far enough to tell.